Mu.Instrumentation.Tracing

Description

This module injects distributed tracing for Mu servers. Currently it only supports Zipkin as backend.

In order to use this module, you need to follow these steps:

Establish a connection with newZipkin.

Wrap the server using zipkin, giving information for the root.

Run the server using the transformer version of your protocol, like |grpcAppTrans|.

Distributed tracing

data MuTracing

Constructors

MuTracing

Fields

samplingPolicy :: SamplingPolicy

rootName :: Text

zipkin :: (MonadIO m, MonadTrace m) => MuTracing -> ServerT chn i p m topHs -> ServerT chn i p m topHs

Wraps a server to do distributed tracing using Zipkin as backend.

runZipkin :: Zipkin -> TraceT m a -> m a

Runs with a given Zipkin connection. You can create one with newZipkin.

Establish connection

newZipkin :: Settings -> IO Zipkin

Create a new connection to Zipkin.

defaultZipkinSettings :: Settings

data Settings

Constructors

Settings

Fields

settingsHostname :: !(Maybe HostName)

settingsPort :: !(Maybe PortNumber)

settingsEndpoint :: !(Maybe Endpoint)

settingsManager :: !(Maybe Manager)

settingsPublishPeriod :: !(Maybe NominalDiffTime)

Useful re-exports